Runs the example workflow correctly
This commit is a quite ugly start, but it's a start.

We correctly copy files to scratch and back again, and run all tasks
correctly. It's now time to schedule jobs on more than one node and
copy files between nodes in a smart way.

After that, it's time for a major cleanup of workflow.py...