Execute targets on remote hosts
This commit uses free cores on any number of hosts to run targets.
Files are automatically moved from host to host, to avoid I/O between
the compute nodes and remote storage.

Next step is to not copy files, but only create hardlinks, if tasks are
run on the same node.

It should be easy to add checkpointing after that.

Closes #4.